refactor: 각 auth 관련 클래스를 패키지별로 관리
