+++ src/main/java/org/ject/support/domain/auth/dto/AuthDto.java
@@ -1,4 +1,4 @@
-package org.ject.support.domain.auth;
+package org.ject.support.domain.auth.dto;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
