+++ src/main/java/org/ject/support/domain/auth/exception/AuthErrorCode.java
@@ -1,4 +1,4 @@
-package org.ject.support.domain.auth;
+package org.ject.support.domain.auth.exception;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
