+++ src/main/java/org/ject/support/domain/auth/exception/AuthException.java
@@ -1,4 +1,4 @@
-package org.ject.support.domain.auth;
+package org.ject.support.domain.auth.exception;
 
 import org.ject.support.common.exception.BusinessException;
 import org.ject.support.common.exception.ErrorCode;
