+++ src/test/java/org/ject/support/domain/auth/AuthControllerTest.java
@@ -8,9 +8,12 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import org.ject.support.common.security.CustomSuccessHandler;
 import org.ject.support.common.security.jwt.JwtTokenProvider;
-import org.ject.support.domain.auth.AuthDto.PinLoginRequest;
-import org.ject.support.domain.auth.AuthDto.TokenRefreshRequest;
-import org.ject.support.domain.auth.AuthDto.VerifyAuthCodeRequest;
+import org.ject.support.domain.auth.dto.AuthDto.PinLoginRequest;
+import org.ject.support.domain.auth.dto.AuthDto.TokenRefreshRequest;
+import org.ject.support.domain.auth.dto.AuthDto.VerifyAuthCodeRequest;
+import org.ject.support.domain.auth.controller.AuthController;
+import org.ject.support.domain.auth.dto.AuthVerificationResult;
+import org.ject.support.domain.auth.service.AuthService;
 import org.ject.support.external.email.domain.EmailTemplate;
 import org.ject.support.testconfig.ApplicationPeriodTest;
 import org.junit.jupiter.api.DisplayName;
