+++ src/test/java/org/ject/support/domain/auth/AuthServiceTest.java
@@ -2,9 +2,9 @@ package org.ject.support.domain.auth;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.ject.support.domain.auth.AuthErrorCode.EXPIRED_REFRESH_TOKEN;
-import static org.ject.support.domain.auth.AuthErrorCode.INVALID_REFRESH_TOKEN;
-import static org.ject.support.domain.auth.AuthErrorCode.INVALID_CREDENTIALS;
+import static org.ject.support.domain.auth.exception.AuthErrorCode.EXPIRED_REFRESH_TOKEN;
+import static org.ject.support.domain.auth.exception.AuthErrorCode.INVALID_REFRESH_TOKEN;
+import static org.ject.support.domain.auth.exception.AuthErrorCode.INVALID_CREDENTIALS;
 import static org.ject.support.domain.member.exception.MemberErrorCode.NOT_FOUND_MEMBER;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
@@ -14,6 +14,9 @@ import static org.mockito.Mockito.mock;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.JwtException;
 import org.ject.support.common.security.jwt.JwtTokenProvider;
+import org.ject.support.domain.auth.dto.AuthVerificationResult;
+import org.ject.support.domain.auth.exception.AuthException;
+import org.ject.support.domain.auth.service.AuthService;
 import org.ject.support.external.email.domain.EmailTemplate;
 import org.ject.support.external.email.exception.EmailErrorCode;
 import org.ject.support.external.email.exception.EmailException;
